pyproject.toml
[project]
name = "x-transcript"
version = "1.0.0"
description = "Transcribe X (Twitter) videos using OpenAI Whisper"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "X-Transcript Team"}
]
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",

    # Queue
    "celery>=5.3.0",
    "redis>=5.0.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # HTTP
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.0",
    "yt-dlp>=2024.0.0",

    # Video/Audio
    "ffmpeg-python>=0.2.0",
    "openai-whisper>=20231117",

    # Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "python-multipart>=0.0.6",
]

[project.scripts]
xtranscript = "app.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
